requirements.txt
slack_sdk>=3.20.0
requests>=2.28.0
neo4j>=5.0.0
prefect>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
Flask>=2.0.0
openai>=1.0.0
flask-cors>=3.0.10
